constructor updates and refactoring for dashboard needs

+++ src/test/java/miltos/diploma/unittests/FxcopAnalyzerTest.java
@@ -88,6 +88,9 @@ public class FxcopAnalyzerTest {
 
     private void clean() throws IOException {
         File output = new File(this.dest);
-        FileUtils.cleanDirectory(output);
+        if (output.exists()) {
+            FileUtils.cleanDirectory(output);
+        }
+        else output.mkdirs();
     }
 }
